requirements.txt
requests
beautifulsoup4
feedparser
python-dateutil
PyYAML
